requirements.txt
asgiref==3.6.0
backports-zoneinfo==0.2.1
certifi==2022.12.7
django-cors-headers==3.14.0
django==4.1.7
djangorestframework==3.14.0
mysqlclient==2.1.1
pip==23.0.1
pytz==2022.7.1
setuptools==65.6.3
sqlparse==0.4.3
tzdata==2022.7
wheel==0.38.4
